requirements/base.txt
django==5.1.6
django-celery-results==2.5.1
celery==5.4.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
psycopg2==2.9.10
redis==5.2.1
ruff==0.11.4
